Fixed the merge into main!
Some adjustments were made including the header and the turn logic upon reaching 100- it turns out a double player move was performed making the game unable to detect a win.

Other than that, everything works except there's a weird bug that detects a double mouse click so the players move twice. But this is rare, and doesn't happen if clicking slowly. Very strange.